fix non-portable format specifiers %zd, %jd, and %zu

+++ src/FSAL/FSAL_POSIX/file.c
@@ -216,8 +216,8 @@ fsal_status_t posix_lock_op (struct fsal_obj_handle * obj_hdl,
         goto out;
     }
     LogFullDebug (COMPONENT_FSAL,
-                  "Locking: op:%d type:%d start:%" PRIu64 " length:%lu ",
-                  lock_op, request_lock->lock_type, request_lock->lock_start, request_lock->lock_length);
+                  "Locking: op:%d type:%d start:%"PRIu64" length:%"PRIu64" ",
+                  (int) lock_op, (int) request_lock->lock_type, (uint64_t) request_lock->lock_start, (uint64_t) request_lock->lock_length);
     if (lock_op == FSAL_OP_LOCKT) {
         fcntl_comm = F_GETLK;
     } else if (lock_op == FSAL_OP_LOCK || lock_op == FSAL_OP_UNLOCK) {
